app/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.5.0
python-dotenv>=1.0.0
opencv-python>=4.8.1.78
Pillow>=10.1.0
numpy>=1.26.0
pandas>=2.1.3

# ML Dependencies - YOLO Ultralytics for segmentation
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
